Eliminate dependency cycle.
