+++ v3/src/models/formula/utils/canonicalization-utils.test.ts
@@ -1,6 +1,6 @@
 import { DisplayNameMap } from "../formula-types"
 import {
-  formulaIndexOf, displayToCanonical, unescapeBacktickString, escapeBacktickString, safeSymbolNameFromDisplayFormula,
+  formulaIndexOf, displayToCanonical, safeSymbolNameFromDisplayFormula,
   makeDisplayNamesSafe, customizeDisplayFormula, canonicalToDisplay
 } from "./canonicalization-utils"
 import { reverseDisplayNameMap, safeSymbolName } from "./name-mapping-utils"
@@ -30,28 +30,6 @@ const displayNameMapExample: DisplayNameMap = {
     }
   }
 }
-
-describe("unescapeBacktickString", () => {
-  it("converts escaped characters in safe symbol name to original characters", () => {
-    expect(unescapeBacktickString("Attribute\\`Test")).toEqual("Attribute`Test")
-    expect(unescapeBacktickString("Attribute\\\\Test")).toEqual("Attribute\\Test")
-  })
-  it("is's an inverse of escapeBacktickString", () => {
-    const testString = "Attribute\\\\\\`Test"
-    expect(unescapeBacktickString(escapeBacktickString(testString))).toEqual(testString)
-  })
-})
-
-describe("escapeBacktickString", () => {
-  it("converts some characters in safe symbol name to escaped characters", () => {
-    expect(escapeBacktickString("Attribute`Test")).toEqual("Attribute\\`Test")
-    expect(escapeBacktickString("Attribute\\Test")).toEqual("Attribute\\\\Test")
-  })
-  it("is's an inverse of unescapeBacktickString", () => {
-    const testString = "Attribute`Test\\"
-    expect(unescapeBacktickString(escapeBacktickString(testString))).toEqual(testString)
-  })
-})
 
 describe("safeSymbolNameFromDisplayFormula", () => {
   it("unescapes special characters and converts strings that are not parsable by Mathjs to valid symbol names", () => {
